test: split sim editor tests along tab seam
test_sim_editor_round_trip.py crossed the 1000-line module cap.
The per-tab widget-state tests (absent-key discipline, per-key
edits, artifact-mode rows, star asymmetry controls) move verbatim
to test_sim_editor_tabs.py; the save/load round-trip acceptance
tests stay.  Same 59 collected tests before and after.
